Move migration + seed + knexfile to src/backend
